feat: Extract combiner AspectVGrid view
