+++ Memorize/Memorize/EmojiMemoryGameView.swift
@@ -27,40 +27,14 @@ struct EmojiMemoryGameView: View {
     }
 
     var cards: some View {
-        GeometryReader { geometry in
-            let gridItemSize = gridItemWidthThatFits(count: viewModel.cards.count, size: geometry.size, atAspectRatio: aspectRatio)
-            LazyVGrid(columns: [GridItem(.adaptive(minimum: gridItemSize), spacing: 0)], spacing: 0) {
-                ForEach(viewModel.cards) { card in
-                    CardView(card)
-                        .aspectRatio(aspectRatio, contentMode: .fit)
-                        .padding(4)
-                        .onTapGesture {
-                            viewModel.choose(card)
-                        }
+        AspectVGrid(viewModel.cards, aspectRatio) { card in
+            CardView(card)
+                .padding(4)
+                .onTapGesture {
+                    viewModel.choose(card)
                 }
-            }
         }
         .foregroundStyle(.orange)
-    }
-
-    func gridItemWidthThatFits(count: Int, size: CGSize, atAspectRatio: CGFloat) -> CGFloat {
-        let count = CGFloat(count)
-        var columnCount = 1.0
-
-        repeat {
-            let width = size.width / columnCount
-            let height = width / atAspectRatio
-
-            let rowColumn = (count / columnCount).rounded(.up)
-
-            if height * rowColumn < size.height {
-                return width.rounded(.down)
-            }
-
-            columnCount += 1
-        } while columnCount < count
-
-        return min(size.width / columnCount, size.height * atAspectRatio).rounded(.down)
     }
 }
